Fix conversion between Date and timestamp/timestamptz types
It seems timestamp values are adjusted for the local time zone.
When converting to a Date, we need to subtract the local offset to get
the correct value.
When converting from a Date, it is better to use the timestamptz type,
which works with both kinds of fields and does not have this problem.