Fix a retain cycle in PostgreSQLConnection (#116)
By adding a handler that retains self on the channel's close future, the connection gets retained until the channel is closed. We break the retain cycle by using a weak reference to self. The completion handler in `send(:onResponse:)` is enough to ensure that the connection stays alive while a message is sent to the server.